LLM_Eval_Gemini/requirements.txt
google-cloud-aiplatform
vertexai
deepeval==3.8.8
python-dotenv==1.2.2
